Add dog struct and functions
This commit adds a new dog struct and functions to initialize, print, create and free a dog. The dog struct has three fields: name (char*), age (float), and owner (char*). The functions included are init_dog(), print_dog(), new_dog(), and free_dog(). This code can be used as a starting point for a program that needs to manage dog information.