Add implementation of sum_them_all function
The sum_them_all function takes an unsigned integer n and a variable number of arguments and returns the sum of all the arguments. If n is 0, the function returns 0. The implementation uses a va_list object to iterate through the arguments and a for loop to add them up.