Add function to add a new node to the beginning of a list
This commit adds a new function, `add_node`, to the `lists.h` header file and implements it in the associated `.c` file. The function takes a pointer to a pointer to the head of a linked list, as well as a string, and adds a new node containing the string to the beginning of the list. The string is duplicated using `strdup`, and its length is calculated using a for loop. If any errors occur during the allocation or duplication process, the function returns NULL.