Implement pop_listint function to delete head node of a linked list
The function takes a pointer to the head node of a linked list as an argument, deletes the head node, and returns the data stored in the deleted node. If the list is empty, the function returns 0. Also, delete the head node by updating the head pointer to point to the second node in the list and freeing the memory allocated for the original head node. The function also uses a temporary pointer to store the address of the original head node, which allows us to access the data value of the deleted node even after it has been freed.