Implement insert_nodeint_at_index function
The function, inserts a new node at a given position in a singly linked list, takes a pointer to a pointer to the first node in the list, an index specifying where to insert the new node, and an integer value to be added to the new node. The function returns a pointer to the new node or NULL if it failed to add the new node.