Implement print_listint_safe function
The function takes in a pointer to the head node of the list and prints the memory address and integer value of each node. It also checks for any loops in the linked list and exits the program with status 98 if a loop is found. Then returns the number of nodes in the list.