Implement clear_bit function
This commit adds a new function called `clear_bit` that sets the value of a bit to 0 at a given index in an unsigned long int. The function takes a pointer to the unsigned long int and the index of the bit to modify as input, and returns 1 if it succeeds or -1 if an error occurs.